Remove hidden dependency on `.env` for public config
The config is now in back-end/config.js and front-end/config.js.
Make sure that both are synchronized to prevent weird bugs.
I tried putting it in the root directory, but React complained about it not being in the src directory and I figured symlinks may break across platforms so it's better to just make a copy and manually keep them in sync.