Add 75s timeout on exec to prevent hanging
The reason why it hangs is not obvious to me. It could be that child_process execFile hangs, the docker client behaves differently in interactive shells vs non-interactive ones, execFile is waiting on stdin for something in ./docker_stop.sh or ./docker_run.sh, I'm not using execFile correctly, pm2 is interfering with execFile, or any combination (including none) of the above.

From a completely blank Docker image state, pulling, stopping, and running took about 50s total. For safety, I added 50% of the time to get to the timeout of 75s. This way we have a working (although hacky) system, and we wait at most 75s between each deployment. However, the containers will be launched before that if some of the layers are cached, so in practice, the user will experience only ~30s of downtime.

These timings were taken on 2021-12-01 on the DigitalOcean vps. They may need adjustment at some time in the future.